fixing pandas X.iloc[X] Dataset getitem method
I was getting a "KeyError" in the 3rd code cell every time in colab and also in VSCode notebooks and finally I could solve it by changing self.X[ix] to self.X.iloc[ix] because they KeyError was from here as pandas wasn't able to get the row item, to get it without the iloc it needs to be at least a slice, not a single row number from what I tested. It is strange because apparently it worked before with the original code but at least for now I had to apply iloc, loc or slicing to get it. BTW, Thanks for your amazing work sensio!!